requirements.txt
click==8.1.3
graphtools==1.5.3
hydra-core==1.3.2
hyperlib==0.0.6
mpmath==1.2.1
networkx==2.8.6
numpy==1.22.0
omegaconf==2.3.0
pandas
phate==1.0.10
PyGSP==0.5.1
python-dotenv==1.0.0
s_gd2==1.8.1
scanpy==1.9.3
scikit_learn==1.2.2
scikit_sparse==0.4.8
scipy==1.9.1
scprep==1.2.1
setuptools==65.3.0
POT
